Created best.R
I created best.R to find the best hospital given a State and one of
three medical outcomes. The function “best” reads in the *.csv of
outcomes, checks the validity of arguments, selects the right data, and
outputs the hospital with the minimum mortality in the selected state
for the selected medical condition.